객체와 메소드의 어색한 연결 수정
member가 상품 목록을 보여주는 것이 아니고 거래 서비스가 상품 목록을 보여준다고 판단해서 TransactionService 클래스를 만들고 해당 메소드를 이동시켰습니다.
